organize database layout into database.go
