+++ modules/transactionpool/persist.go
@@ -1,7 +1,6 @@
 package transactionpool
 
 import (
-	"encoding/json"
 	"errors"
 	"fmt"
 	"os"
@@ -9,7 +8,6 @@ import (
 	"time"
 
 	"github.com/NebulousLabs/Sia/build"
-	"github.com/NebulousLabs/Sia/encoding"
 	"github.com/NebulousLabs/Sia/modules"
 	"github.com/NebulousLabs/Sia/persist"
 	"github.com/NebulousLabs/Sia/types"
@@ -20,46 +18,15 @@ import (
 const tpoolSyncRate = time.Minute * 2
 
 var (
-	// bucketRecentConsensusChange holds the most recent consensus change seen
-	// by the transaction pool.
-	bucketRecentConsensusChange = []byte("RecentConsensusChange")
-
-	// bucketBlockHeight holds the most recent block height seen by the
-	// transaction pool.
-	bucketBlockHeight = []byte("BlockHeight")
-
-	// bucketConfirmedTransactions holds the ids of every transaction that has
-	// been confirmed on the blockchain.
-	bucketConfirmedTransactions = []byte("ConfirmedTransactions")
-
-	// bucketFeeMedian stores all of the persist data relating to the fee
-	// median.
-	bucketFeeMedian = []byte("FeeMedian")
-
 	// errNilConsensusChange is returned if there is no consensus change in the
 	// database.
 	errNilConsensusChange = errors.New("no consensus change found")
 
-	// fieldRecentConsensusChange is the field in bucketRecentConsensusChange
-	// that holds the value of the most recent consensus change.
-	fieldRecentConsensusChange = []byte("RecentConsensusChange")
-
-	// fieldBlockHeight is the field in bucketBlockHeight that holds the value of
-	// the most recent block height.
-	fieldBlockHeight = []byte("BlockHeight")
-
-	// fieldFeeMedian is the fee median persist data stored in a fee median
-	// field.
-	fieldFeeMedian = []byte("FeeMedian")
+	// errNilFeeMedian is the message returned if a database does not find fee
+	// median persistance.
+	errNilFeeMedian = errors.New("no fee median found")
 )
 
-// feeMedian is the persist format for the fee median data storage.
-type feeMedian struct {
-	BlockHeight         types.BlockHeight
-	RecentConfirmedFees []feeSummary
-	TxnsPerBlock        []uint64
-	RecentMedianFee     types.Currency
-}
 
 // threadedRegularSync will make sure that sync gets called on the database
 // every once in a while.
@@ -196,19 +163,17 @@ func (tp *TransactionPool) initPersist() error {
 	}
 
 	// Get the fee median data.
-	var mp medianPersist
-	medianBytes := tp.dbTx.Bucket(bucketFeeMedian).Get(fieldFeeMedian)
-	// If medianBytes is 'nil', it just means that median data has never been
-	// saved.
-	if medianBytes != nil {
-		err = json.Unmarshal(medianBytes, &mp)
-		if err != nil {
-			return build.ExtendErr("unable to unmarshal median data:", err)
-		}
+	mp, err := tp.getFeeMedian(tp.dbTx)
+	if err != nil && err != errNilFeeMedian {
+		return build.ExtendErr("unable to load the fee median", err)
 	}
-	tp.recentConfirmedFees = mp.RecentConfirmedFees
-	tp.txnsPerBlock = mp.TxnsPerBlock
-	tp.recentMedianFee = mp.RecentMedianFee
+	// Just leave the fields empty if no fee median was found. They will be
+	// filled out.
+	if err != errNilFeeMedian {
+		tp.recentConfirmedFees = mp.RecentConfirmedFees
+		tp.txnsPerBlock = mp.TxnsPerBlock
+		tp.recentMedianFee = mp.RecentMedianFee
+	}
 
 	// Subscribe to the consensus set using the most recent consensus change.
 	err = tp.consensusSet.ConsensusSetSubscribe(tp, cc)
@@ -237,44 +202,6 @@ func (tp *TransactionPool) initPersist() error {
 	return nil
 }
 
-// getBlockHeight returns the most recent block height from the database.
-func (tp *TransactionPool) getBlockHeight(tx *bolt.Tx) (bh types.BlockHeight, err error) {
-	err = encoding.Unmarshal(tx.Bucket(bucketBlockHeight).Get(fieldBlockHeight), &bh)
-	return
-}
-
-// putBlockHeight updates the transaction pool's block height.
-func (tp *TransactionPool) putBlockHeight(tx *bolt.Tx, height types.BlockHeight) error {
-	tp.blockHeight = height
-	return tx.Bucket(bucketBlockHeight).Put(fieldBlockHeight, encoding.Marshal(height))
-}
-
-// putMedianFees puts a median fees object into the database.
-func (tp *TransactionPool) putMedianFees(tx *bolt.Tx, mp medianPersist) error {
-	objBytes, err := json.Marshal(mp)
-	if err != nil {
-		return err
-	}
-	return tx.Bucket(bucketFeeMedian).Put(fieldFeeMedian, objBytes)
-}
-
-// getRecentConsensusChange returns the most recent consensus change from the
-// database.
-func (tp *TransactionPool) getRecentConsensusChange(tx *bolt.Tx) (cc modules.ConsensusChangeID, err error) {
-	ccBytes := tx.Bucket(bucketRecentConsensusChange).Get(fieldRecentConsensusChange)
-	if ccBytes == nil {
-		return modules.ConsensusChangeID{}, errNilConsensusChange
-	}
-	copy(cc[:], ccBytes)
-	return cc, nil
-}
-
-// putRecentConsensusChange updates the most recent consensus change seen by
-// the transaction pool.
-func (tp *TransactionPool) putRecentConsensusChange(tx *bolt.Tx, cc modules.ConsensusChangeID) error {
-	return tx.Bucket(bucketRecentConsensusChange).Put(fieldRecentConsensusChange, cc[:])
-}
-
 // transactionConfirmed returns true if the transaction has been confirmed on
 // the blockchain and false if the transaction has not been confirmed on the
 // blockchain.
@@ -284,15 +211,4 @@ func (tp *TransactionPool) transactionConfirmed(tx *bolt.Tx, id types.Transactio
 		return false
 	}
 	return true
-}
-
-// addTransaction adds a transaction to the list of confirmed transactions.
-func (tp *TransactionPool) addTransaction(tx *bolt.Tx, id types.TransactionID) error {
-	return tx.Bucket(bucketConfirmedTransactions).Put(id[:], []byte{})
-}
-
-// deleteTransaction deletes a transaction from the list of confirmed
-// transactions.
-func (tp *TransactionPool) deleteTransaction(tx *bolt.Tx, id types.TransactionID) error {
-	return tx.Bucket(bucketConfirmedTransactions).Delete(id[:])
 }
