+++ modules/transactionpool/transactionpool.go
@@ -36,20 +36,6 @@ type (
 	ObjectID         crypto.Hash
 	TransactionSetID crypto.Hash
 
-	// feeSummary desribes the fee structure of a transaction.
-	feeSummary struct {
-		Fee  types.Currency // SC per byte
-		Size uint64
-	}
-
-	// medianPersist is the json object that gets stored in the database so that
-	// the transaction pool can persist its block based fee estimations.
-	medianPersist struct {
-		RecentConfirmedFees []feeSummary
-		TxnsPerBlock        []uint64
-		RecentMedianFee     types.Currency
-	}
-
 	// The TransactionPool tracks incoming transactions, accepting them or
 	// rejecting them based on internal criteria such as fees and unconfirmed
 	// double spends.
